+++ common-custom-user-data-gradle-plugin/src/main/java/com/gradle/CustomBuildScanEnhancements.java
@@ -7,21 +7,10 @@ import org.gradle.api.Task;
 import org.gradle.api.invocation.Gradle;
 import org.gradle.api.tasks.testing.Test;
 
-import java.util.Properties;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-import static com.gradle.Utils.appendIfMissing;
-import static com.gradle.Utils.envVariable;
-import static com.gradle.Utils.envVariablePresent;
-import static com.gradle.Utils.execAndCheckSuccess;
-import static com.gradle.Utils.execAndGetStdOut;
-import static com.gradle.Utils.isNotEmpty;
-import static com.gradle.Utils.readPropertiesFile;
-import static com.gradle.Utils.sysProperty;
-import static com.gradle.Utils.sysPropertyKeyStartingWith;
-import static com.gradle.Utils.sysPropertyPresent;
-import static com.gradle.Utils.urlEncode;
+import static com.gradle.Utils.*;
 
 /**
  * Adds a standard set of useful tags, links and custom values to all build scans published.
@@ -83,20 +72,14 @@ final class CustomBuildScanEnhancements {
         }
 
         if (isTeamCity()) {
-            if (sysPropertyPresent("teamcity.configuration.properties.file")) {
-                Properties properties = readPropertiesFile(sysProperty("teamcity.configuration.properties.file"));
-                String teamCityServerUrl = properties.getProperty("teamcity.serverUrl");
-                if (teamCityServerUrl != null && sysPropertyPresent("build.number") && sysPropertyPresent("teamcity.buildType.id")) {
-                    String buildNumber = sysProperty("build.number");
-                    String buildTypeId = sysProperty("teamcity.buildType.id");
-                    buildScan.link("TeamCity build", appendIfMissing(teamCityServerUrl, "/") + "viewLog.html?buildNumber=" + buildNumber + "&buildTypeId=" + buildTypeId);
-                }
+            if (envVariablePresent("BUILD_URL")) {
+                buildScan.link("TeamCity build", envVariable("BUILD_URL"));
             }
-            if (sysPropertyPresent("build.number")) {
-                buildScan.value("CI build number", sysProperty("build.number"));
+            if (envVariablePresent("BUILD_NUMBER")) {
+                buildScan.value("CI build number", envVariable("BUILD_NUMBER"));
             }
-            if (sysPropertyPresent("agent.name")) {
-                addCustomValueAndSearchLink(buildScan, "CI agent", sysProperty("agent.name"));
+            if (envVariablePresent("BUILD_AGENT_NAME")) {
+                addCustomValueAndSearchLink(buildScan, "CI agent", envVariable("BUILD_AGENT_NAME"));
             }
         }